reorganize footing problem
